Refactor namespaces and update dependencies
Updated `using` directives and namespaces in `CommandFailureBehavior.cs` and `RegisterUserUserCase.cs` to reflect new project structure. Modified `AuthController` to include new dependencies and added a `Register` method with Swagger annotations for user registration.
